chore(auth): extract JWT auth guard to shared package (#1161)
Move nearly identical JWT authentication guard implementations from 6
services (user, discussion, notification, ai, contact, moderation)
to @reason-bridge/common package.

Changes:
- Add JwtAuthGuard and JwtPayload to packages/common/src/auth/
- Add peer dependencies for @nestjs/jwt, @nestjs/config, jwks-rsa
- Enable experimentalDecorators and emitDecoratorMetadata in common tsconfig
- Replace service implementations with re-exports for backward compatibility

The shared guard supports:
- Mock/development/test mode (HS256 with JWT_SECRET)
- Production mode (RS256 with Cognito JWKS)
- Automatic mode detection based on environment variables

Reduces ~900 lines of duplicated code to a single 170-line implementation.

Closes #1161
