+++ packages/common/src/index.ts
@@ -15,3 +15,4 @@ export * from './events/index.js';
 export * from './validation/index.js';
 export * from './config/index.js';
 export * from './observability/index.js';
+export * from './auth/index.js';
